initial commit
This is a big file inspector I wrote this python concept and cleaned it up with the help of ai when i was having issues with the dependencies so i then had the help of ai to make it run with limited basic dependencies it now runs off of stock kali linux and i would assume any linux with basic dependencies. Great for searching large database .txt files and that was why i made it in the first place after crashing my computer so many times trying to open a 16gb .txt file yes im new to the data searches lol.